Streaming with less latency
I suggest to use the switch "--live" to stream video with less latency. After this i can see my cam realtime.
I suggest to use the switch "--time" to be more tollerant about loosing connection. In this way is more stable. 
I remove the "I" swirch because i dont want to show the streaming information and is more fast to load.